requirements.txt
tqdm
trimesh
lightning
matplotlib
robust_laplacian
libigl
tensorboard
potpourri3d
pymeshlab
rich
websockets
line_profiler
joblib
pytorch_lightning
trimesh
matplotlib
colorcet
libigl
tensorboard
pygeodesic
plotly